+++ src/OFS/ObjectManager.py
@@ -274,6 +274,13 @@ class ObjectManager(
                 if container_filter(self):
                     meta_types.append(entry)
 
+        # Synthesize a new key into each item that decides whether to show
+        # the modal add dialog in the Zope 4 ZMI
+        for mt in meta_types:
+            want_modal = getattr(mt.get('instance', None),
+                                 'zmi_show_add_dialog', True)
+            mt['zmi_show_add_dialog'] = 'modal' if want_modal else ''
+
         return meta_types
 
     def _subobject_permissions(self):